config_sqlite.py
import os
from dotenv import load_dotenv

load_dotenv()

class Config:
    SQLITE_DB = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'kawan_umkm.db')
    SECRET_KEY = os.getenv('SECRET_KEY', 'your-secret-key-here')
    JWT_SECRET_KEY = os.getenv('JWT_SECRET_KEY', 'jwt-secret-key-here')
    UPLOAD_FOLDER = os.path.join(os.path.dirname(os.path.abspath(__file__)), 'uploads', 'images')
    MAX_CONTENT_LENGTH = 100 * 1024 * 1024
    
    ALLOWED_EXTENSIONS = {'png', 'jpg', 'jpeg', 'gif', 'webp'}

    @staticmethod
    def init_app(app):
        os.makedirs(Config.UPLOAD_FOLDER, exist_ok=True)