requirements.txt
Flask==2.3.3
Flask-MySQLdb==1.0.1
Flask-JWT-Extended==4.5.3
Flask-CORS==4.0.0
Werkzeug==2.3.7
Pillow==10.0.1
python-dotenv==1.0.0
bcrypt==4.0.1
PyJWT==2.8.0